massmailer/requirements.txt
# requirements.txt
Flask==2.3.3
Flask-Mail==0.9.1
Flask-WTF==1.0.1
Jinja2==3.1.6
MarkupSafe==3.0.2
python-dotenv==1.0.0
pandas==2.1.3
openpyxl==3.1.2
WTForms==3.0.1
itsdangerous==2.1.2
email-validator==2.1.0
requests==2.31.0
html5lib==1.1
cssselect==1.2.0
chardet==5.2.0
idna==3.4
urllib3==1.26.16
charset-normalizer==2.1.1
certifi==2023.7.22
six==1.16.0
pytz==2023.3.post1
python-dateutil==2.8.2
langchain_openai
langchain_core
langchain_community
python-dotenv
streamlit
langserve
fastapi
uvicorn
